requirements.txt
fastapi==0.111.0
# Drop extras to reduce optional dependencies
uvicorn==0.29.0
gunicorn==21.2.0
sqlalchemy==2.0.36
asyncpg==0.29.0

psycopg[binary]==3.2.*
alembic==1.13.1
pydantic-settings==2.0.3
pandas==2.2.3
numpy==2.3.2
yfinance==0.2.65
typer==0.16.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.12.0
pyyaml==6.0.1
aiosqlite==0.21.0
trio==0.30.0
aiofiles==23.2.1
python-multipart==0.0.6
